Moved daily log and help menu functions to their own file + some error handling and bug fixes
