+++ Hustle-Hamster/menuHelp.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <iostream>
+
+#include "utils.h"
+#include "dailyLog.h"
+
+/**
+ * Method to display the help menu
+ * Will present users with options and the user will select one
+ * The method will then direct user to the selected option
+*/
+void helpMenu();
+
+/**
+ * Prints a tutorial for the user
+*/
+void howToUse();
